- Added the Customers This was originally an object filling all the properties with inheritance which BasicCustomer and VIPCustomer were pointing to a bass class Customer.
However, because I am going to use DOD, the classes
are replaced by two hashset (technically still
hashmap because hashset is not available in zig)
representing basic customers and VIP customers,
and replacing the optional discount properties
with another hashmap.

This design will remove the null field or the
need of classes, and each chuck of data will have
smaller memory footprint to avoid cache misses
because the program can now load the customer data
partially instead of fetching the whole object
with unused properties, while the two set is clear
enough to separately the difference between VIP
and basic customers.