- Added Fade in and out animation My original assignment flexed with a demoscene, with the title fading in and out with a bit of scrolling, and I decided to bring it to the zig version as well.
The reason is that I wanna know more about how
strings are handled in zig where I cheat in the
other projects which I have used string dependency.
Without messing around with the raw []u8 and how
to passing those as a function parameters, it is
likely to face some hurtles when I working on some
more complex future projects involving strings.